% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{addPlotOptions}
\alias{addPlotOptions}
\title{Adds plot options to the current previewItem string.}
\usage{
addPlotOptions(previewItemString, showCI, stackPlots)
}
\arguments{
\item{previewItemString}{string representing the previewItem}

\item{showCI}{if confidence interval should be shown}

\item{stackPlots}{if subplots should be stacked (on top of each other) or not}
}
\value{
the updated preview item string
}
\description{
Adds plot options to the current previewItem string.
}
