% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{addPreviewItemRibbon}
\alias{addPreviewItemRibbon}
\title{Adds the given ribbon to the current previewItem string.}
\usage{
addPreviewItemRibbon(previewItemString, ribbon)
}
\arguments{
\item{previewItemString}{string representing the previewItem}

\item{ribbon}{ribbon value}
}
\value{
the updated preview item string
}
\description{
Adds the given ribbon to the current previewItem string.
}
