% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{addPreviewItemType}
\alias{addPreviewItemType}
\title{Adds the given type to the current previewItem string.}
\usage{
addPreviewItemType(previewItemString, previewItemType)
}
\arguments{
\item{previewItemString}{string representing the previewItem}

\item{previewItemType}{preview item type}
}
\value{
the updated preview item string
}
\description{
Adds the given type to the current previewItem string.
}
