% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{addPreviewItemTypeAge}
\alias{addPreviewItemTypeAge}
\title{Adds the given type to the current previewItem string, by sex}
\usage{
addPreviewItemTypeAge(previewItemString, previewItemType)
}
\arguments{
\item{previewItemString}{string representing the previewItem}

\item{previewItemType}{preview item type}
}
\value{
the updated preview item string
}
\description{
Adds the given type to the current previewItem string, by sex
}
