% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataExtraction.R
\name{bloodCancerPrevalenceExtraction}
\alias{bloodCancerPrevalenceExtraction}
\title{Extracts denominator data from files in CSV format.}
\usage{
bloodCancerPrevalenceExtraction()
}
\value{
A tibble
}
\description{
Extracts denominator data from files in CSV format.
}
