% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataExtraction.R
\name{bloodCancerPrevalenceExtractionToRDS}
\alias{bloodCancerPrevalenceExtractionToRDS}
\title{Extracts denominator data from files in CSV format.}
\usage{
bloodCancerPrevalenceExtractionToRDS()
}
\value{
A tibble
}
\description{
Extracts denominator data from files in CSV format.
}
