% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createSunburstPlot.R
\name{buildHierarchy}
\alias{buildHierarchy}
\title{Creates a hierarchical data structure.}
\usage{
buildHierarchy(csv)
}
\arguments{
\item{csv}{CSV}
}
\value{
JSON
}
\description{
Creates a hierarchical data structure.
}
