% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treatmentPatterns.R
\name{createLegend}
\alias{createLegend}
\title{Write html file for the sunburst plot legend.}
\usage{
createLegend(inputJSON, fileName)
}
\arguments{
\item{inputJSON}{JSON-data to create legend}

\item{fileName}{Output file name for the legend.}
}
\description{
Write html file for the sunburst plot legend.
}
