% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataExtraction.R
\name{dataCleanAttrition}
\alias{dataCleanAttrition}
\title{Clean attrition data}
\usage{
dataCleanAttrition(attrition)
}
\arguments{
\item{attrition}{attrition data}
}
\value{
the updated attrition
}
\description{
`dataCleanAttrition()` clean incidence/prevalence attrition data
}
