% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataExtraction.R
\name{denominatorData}
\alias{denominatorData}
\title{Extracts denominator data from files in CSV format.}
\usage{
denominatorData()
}
\value{
A tibble
}
\description{
Extracts denominator data from files in CSV format.
}
