% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataExtraction.R
\name{denominatorExtraction}
\alias{denominatorExtraction}
\title{Extracts data from csv files}
\usage{
denominatorExtraction(
  importFolderDenominator = here("inst/csv/denominatorMockData")
)
}
\arguments{
\item{importFolderDenominator}{Location of csv files.}
}
\value{
A tibble.
}
\description{
Extracts data from csv files
}
