% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataExtraction.R
\name{extractCSV}
\alias{extractCSV}
\title{`extractCSV()` iterates every file in the databaseFolders to `processCSV()`}
\usage{
extractCSV(databaseFolders, configData)
}
\arguments{
\item{databaseFolders}{A list of full name folder locations}

\item{configData}{Configuration from yaml file}
}
\value{
A list objects with summarisedResults
}
\description{
`extractCSV()` iterates every file in the databaseFolders to `processCSV()`
}
