% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateMockData.R
\name{generateMockData}
\alias{generateMockData}
\title{Creates a mock data set for ReportGenerator}
\usage{
generateMockData(
  databaseName = c("CHUBX", "CPRD GOLD", "IMASIS", "IPCI", "SIDIAP"),
  simulatePopulation = TRUE,
  sampleSize = NULL,
  outputDir = file.path(getwd(), "results")
)
}
\arguments{
\item{databaseName}{A vector with the name in characters of each database.}

\item{simulatePopulation}{TRUE or FALSE to simulate different population sizes. TRUE is default.}

\item{sampleSize}{Size of the sample for each database.}

\item{outputDir}{A character vector of the directory to export mock data.}
}
\value{
csv files
}
\description{
`generateMockData()` uses mock functions from other packages to create csv files to test ReportGenerator.
}
