% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate.R
\name{generatePresentation}
\alias{generatePresentation}
\title{generatePresentation}
\usage{
generatePresentation(
  server,
  username,
  password,
  dbms,
  resultsSchema = NULL,
  targetId,
  subsetId = NULL,
  outcomeId,
  comparatorId,
  covariateIds = NULL,
  friendlyNames = list(targetName = "target cohort", comparatorName =
    "comparator cohort", indicationName = "indication cohort", outcomeName =
    "outcome name"),
  details = list(studyPeriod = "All Time", restrictions = "Age - None"),
  title = "ASSURE 001 ...",
  lead = "add name",
  date = Sys.Date(),
  backgroundText = "",
  evaluationText = "",
  outputLocation = getwd(),
  outputName = paste0("presentation_", gsub(":", "_", gsub(" ", "_",
    as.character(date()))), ".html"),
  intermediateDir = tempdir()
)
}
\arguments{
\item{server}{The server containing the result database}

\item{username}{The username for an account that can access the result database}

\item{password}{The password for an account that can access the result database}

\item{dbms}{The dbms used to access the result database}

\item{resultsSchema}{The result database schema}

\item{targetId}{The cohort definition id for the target cohort}

\item{subsetId}{Optional a subset ID for the cohort method/characterization and prediction results}

\item{outcomeId}{The cohort definition id for the outcome}

\item{comparatorId}{The cohort method comparator cohort id}

\item{covariateIds}{A vector of covariateIds to include in the characterization}

\item{friendlyNames}{a data.frame with friendly name conversions}

\item{details}{a list with the studyPeriod and restrictions}

\item{title}{A title for the presentation}

\item{lead}{The name of the presentor}

\item{date}{The date of the presentation}

\item{backgroundText}{a character with any background text}

\item{evaluationText}{a list of bullet points for the evaluation}

\item{outputLocation}{The file location and name to save the protocol}

\item{outputName}{The name of the html protocol that is created}

\item{intermediateDir}{The work directory for quarto}
}
\value{
An named R list with the elements 'standard' and 'source'
}
\description{
Generates a presentation from a Strategus result
}
\details{
Specify the connection details to the result database and the schema name
to generate a presentation.
}
