% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Helpers.R
\name{getCharacterization}
\alias{getCharacterization}
\title{A function to extract TnO and TnOc characterization results}
\usage{
getCharacterization(
  server,
  username,
  password,
  dbms,
  schema,
  targetId,
  outcomeId,
  covariateIds
)
}
\arguments{
\item{server}{the server of the results}

\item{username}{the connection username to the result database}

\item{password}{the connection password to the result database}

\item{dbms}{the connection dbms to the result database}

\item{schema}{the result database schema}

\item{targetId}{The cohort id for the target cohort}

\item{outcomeId}{The cohort id for the outcome cohort}

\item{covariateIds}{The vector of covariateIds to extract}
}
\value{
A data.frame with the specified characterization results
}
\description{
A function to extract TnO and TnOc characterization results
}
\details{
The user specifies results database and schema and the covariateIds of interest
}
