% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{getItemConfig}
\alias{getItemConfig}
\title{Checks items available for ReportGenerator}
\usage{
getItemConfig(input = NULL, output = NULL, inputValue = NULL)
}
\arguments{
\item{input}{Declare the type of value such as "title" to look up in the yml file.}

\item{output}{Expected value from the yml file such as the "function" to evaluate.}

\item{inputValue}{The actual value in character to look up.}
}
\value{
a dataframe with the properties of the items
}
\description{
Get the items that the user can choose from in the report generator. The list is loaded from the configuration file
and filtered by the files that have been uploaded.
}
