% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{getItemOptions}
\alias{getItemOptions}
\title{Get options for given item.}
\usage{
getItemOptions(title)
}
\arguments{
\item{item}{the menu item}
}
\value{
the options
}
\description{
Get options for given item.
}
