% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportGenerator.R
\name{getItemsList}
\alias{getItemsList}
\title{Get the items that the user can choose from in the report generator. The list is loaded from the configuration file
and filtered by the files that have been uploaded.}
\usage{
getItemsList(uploadedFiles)
}
\arguments{
\item{uploadedFiles}{vector of uploaded filenames.}
}
\value{
a dataframe with the properties of the items
}
\description{
Get the items that the user can choose from in the report generator. The list is loaded from the configuration file
and filtered by the files that have been uploaded.
}
