% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataExtraction.R
\name{incidenceData}
\alias{incidenceData}
\title{Extracts incidence data from files in CSV format.}
\usage{
incidenceData()
}
\value{
A tibble
}
\description{
Extracts incidence data from files in CSV format.
}
