% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataExtraction.R
\name{incidenceExtraction}
\alias{incidenceExtraction}
\title{Extracts data from csv files and saves rds}
\usage{
incidenceExtraction(
  importFolderIndcidence = here("inst/csv/incidenceMockResults"),
  studyName = "mock_data"
)
}
\arguments{
\item{importFolderIndcidence}{Location of csv files.}

\item{studyName}{Name of the study.}
}
\value{
A tibble
}
\description{
Extracts data from csv files and saves rds
}
