% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/incidenceFigures.R
\name{incidenceFigure1}
\alias{incidenceFigure1}
\title{Figure 1. Incidence rate/s of drug/s use over calendar incidence_start_date (per month/year) overall}
\usage{
incidenceFigure1(incidenceData)
}
\arguments{
\item{incidenceData}{}
}
\value{
incidenceFigure1
}
\description{
Figure 1. Incidence rate/s of drug/s use over calendar incidence_start_date (per month/year) overall
}
