% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/incidenceFigures.R
\name{incidenceFigure2}
\alias{incidenceFigure2}
\title{Figure 2. by year/month: two plots – males/females, all databases}
\usage{
incidenceFigure2(incidenceData)
}
\arguments{
\item{incidenceData}{}
}
\value{
incidenceFigure2
}
\description{
Figure 2. by year/month: two plots – males/females, all databases
}
