% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/incidenceFigures.R
\name{incidenceFigure3}
\alias{incidenceFigure3}
\title{Figure 3 . by year/month: plot for each database, diff lines per age group}
\usage{
incidenceFigure3(incidenceData)
}
\arguments{
\item{incidenceData}{}
}
\value{
incidenceFigure3
}
\description{
Figure 3 . by year/month: plot for each database, diff lines per age group
}
