% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/incidenceFigures.R
\name{incidenceFigure4}
\alias{incidenceFigure4}
\title{Figure 4 . by age group (x-axis) for databases (color) and sex (dashed/line)}
\usage{
incidenceFigure4(incidenceData)
}
\arguments{
\item{indcidenceData}{}
}
\value{
incidenceFigure4
}
\description{
Figure 4 . by age group (x-axis) for databases (color) and sex (dashed/line)
}
