% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/incidencePrevalenceReport.R
\name{incidencePrevalenceReport}
\alias{incidencePrevalenceReport}
\title{Creates an HTML report taking the prevalence and incidence estimates from IncidencePrevalence.}
\usage{
incidencePrevalenceReport(
  studyTitle,
  studyAuthor,
  abstractText,
  denominatorData,
  incidenceData,
  prevalenceData,
  format = "word"
)
}
\arguments{
\item{studyTitle}{Title of the study.}

\item{studyAuthor}{Name of the author.}

\item{abstractText}{Abstract in character.}

\item{denominatorData}{A tibble from IncidencePrevalence.}

\item{incidenceData}{A tibble from IncidencePrevalence.}

\item{prevalenceData}{A tibble from IncidencePrevalence.}

\item{format}{Output options are "word", "pdf" or "html".}
}
\value{
A WORD, PDF or HTML document.
}
\description{
Creates an HTML report taking the prevalence and incidence estimates from IncidencePrevalence.
}
