% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardPlots.R
\name{incidenceRatePerYearColorByAgePlot}
\alias{incidenceRatePerYearColorByAgePlot}
\title{Incidence Rate Per Year by Age Plot}
\usage{
incidenceRatePerYearColorByAgePlot(incidence_estimates, type)
}
\arguments{
\item{incidence_estimates}{estimates of of the incidence}

\item{type}{plot type}
}
\description{
Incidence Rate Per Year by Age Plot
}
