% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardPlots.R
\name{incidenceRatePerYearGroupBySexPlot}
\alias{incidenceRatePerYearGroupBySexPlot}
\title{incidenceRatePerYearGroupBySexPlot}
\usage{
incidenceRatePerYearGroupBySexPlot(incidence_estimates)
}
\arguments{
\item{incidence_estimates}{estimates of of the incidence}
}
\description{
incidenceRatePerYearGroupBySexPlot
}
