% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardPlots.R
\name{incidenceRatePerYearPlot}
\alias{incidenceRatePerYearPlot}
\title{Incidence Rate Per Year Plot}
\usage{
incidenceRatePerYearPlot(incidence_estimates, type)
}
\arguments{
\item{incidence_estimates}{estimates of of the incidence}

\item{plot}{type}
}
\description{
Incidence Rate Per Year Plot
}
