% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/incidenceTables.R
\name{incidenceTable1}
\alias{incidenceTable1}
\title{Table 1. <Drug users / Patients with condition X> in <list all data sources> during <stud period>}
\usage{
incidenceTable1(incidenceData = NULL)
}
\arguments{
\item{incidenceData}{}
}
\value{
incidenceTable1
}
\description{
Table 1. <Drug users / Patients with condition X> in <list all data sources> during <stud period>
}
