% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{increaseFacetStripSize}
\alias{increaseFacetStripSize}
\title{Increase the facet strip size for better readability.}
\usage{
increaseFacetStripSize(plotlyObject)
}
\arguments{
\item{plotlyObject}{object generated by ggplotly()}
}
\value{
the updated object
}
\description{
Increase the facet strip size for better readability.
}
