% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataExtraction.R
\name{joinDatabase}
\alias{joinDatabase}
\title{`joinDatabase()` joins several zip or csv folders into a list of dataframes.}
\usage{
joinDatabase(
  fileDataPath = NULL,
  fileName = NULL,
  package = "IncidencePrevalence",
  versionData = "0.5.1",
  csvLocation = NULL
)
}
\arguments{
\item{fileDataPath}{File path(s) in character}

\item{fileName}{Name of the file in character to process in case the input is only csv}

\item{package}{Name of the packages that generated the results}

\item{versionData}{Version of the package}

\item{csvLocation}{Path folder location to uncompress the zip files}
}
\value{
A list of dataframes
}
\description{
`joinDatabase()` joins several zip or csv folders into a list of dataframes.
}
