% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataExtraction.R
\name{joinDatabases}
\alias{joinDatabases}
\title{`joinDatabases()` joins several zip or csv folders into a list of dataframes.}
\usage{
joinDatabases(fileDataPath, fileName = NULL, logger)
}
\arguments{
\item{fileDataPath}{List of full file locations}

\item{fileName}{Name of the file in character to process in case the input is only csv}

\item{logger}{A logger object}
}
\value{
A list of dataframes
}
\description{
`joinDatabases()` joins several zip or csv folders into a list of dataframes.
}
