% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataExtraction.R
\name{joinZipFiles}
\alias{joinZipFiles}
\title{Joins zip databases}
\usage{
joinZipFiles(uploadedFiles = NULL)
}
\arguments{
\item{filesLocation}{Location of zip files}
}
\value{
A list of files
}
\description{
`joinZipFiles()` joins data from multiple datapartners and returns a list of condensed files in a temp dir
}
