% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataExtraction.R
\name{loadFileData}
\alias{loadFileData}
\title{`loadFileData()` checks every csv their result type using columns or summarisedResult class as reference}
\usage{
loadFileData(
  data,
  fileName,
  configData,
  resultsData,
  resultsColumns,
  databaseName
)
}
\arguments{
\item{data}{named list with data.}

\item{fileName}{the file to load}

\item{configData}{config loaded from file}

\item{resultsData}{the loaded file data}

\item{resultsColumns}{the loaded file columns}

\item{databaseName}{db name}
}
\value{
the list with data
}
\description{
`loadFileData()` checks every csv their result type using columns or summarisedResult class as reference
}
