% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mockSampleCSV.R
\name{mockSampleCSV}
\alias{mockSampleCSV}
\title{This function uses data mockIncidencePrevalenceRef function to simulate three databases in CSV format to test the ReportGenerator}
\usage{
mockSampleCSV(databaseName = c("Synthea", "IPCI", "CPRD"))
}
\arguments{
\item{databaseName}{A vector that specifies the name of each database.}
}
\value{
CSV files inside inst/csv
}
\description{
This function uses data mockIncidencePrevalenceRef function to simulate three databases in CSV format to test the ReportGenerator
}
