% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createSunburstPlot.R
\name{prepData}
\alias{prepData}
\title{prepData}
\usage{
prepData(treatmentHistory, year)
}
\arguments{
\item{treatmentHistory}{(`data.frame()`)}

\item{year}{(`integer(1)`)}
}
\value{
(`data.frame()`)
}
\description{
prepData
}
