% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataExtraction.R
\name{prevalenceData}
\alias{prevalenceData}
\title{Extracts prevalence data from files in CSV format.}
\usage{
prevalenceData()
}
\value{
A tibble
}
\description{
Extracts prevalence data from files in CSV format.
}
