% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataExtraction.R
\name{prevalenceExtraction}
\alias{prevalenceExtraction}
\title{Extracts data from csv and saves rds}
\usage{
prevalenceExtraction(
  importFolderPrevalence = here("inst/csv/prevalenceMockResults"),
  studyName = "mock_data"
)
}
\arguments{
\item{importFolderPrevalence}{Location of prevalence results.}

\item{studyName}{Name of the study.}
}
\value{
A tibble
}
\description{
Extracts data from csv and saves rds
}
