% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataExtraction.R
\name{prevalenceExtractionToRDS}
\alias{prevalenceExtractionToRDS}
\title{Extracts data from csv files and saves rds}
\usage{
prevalenceExtractionToRDS(
  importFolderPrevalence = here("inst/csv/prevalenceMockResults"),
  studyName = "mock_data"
)
}
\arguments{
\item{importFolderPrevalence}{Location of prevalence results.}

\item{studyName}{Name of the study.}
}
\value{
A tibble
}
\description{
Extracts data from csv files and saves rds
}
