% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prevalenceFigures.R
\name{prevalenceFigure5}
\alias{prevalenceFigure5}
\title{Figure 5. Prevalence of drug/s use over calendar time (per month/year) overall}
\usage{
prevalenceFigure5(prevalenceData)
}
\arguments{
\item{prevalenceData}{}
}
\value{
prevalenceFigureData
}
\description{
Figure 5. Prevalence of drug/s use over calendar time (per month/year) overall
}
