% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prevalenceFigures.R
\name{prevalenceFigure6}
\alias{prevalenceFigure6}
\title{Figure 6. by year/month: two plots – males/females, all databases}
\usage{
prevalenceFigure6(prevalenceData)
}
\arguments{
\item{prevalenceData}{}
}
\value{
prevalenceFigure6
}
\description{
Figure 6. by year/month: two plots – males/females, all databases
}
