% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prevalenceFigures.R
\name{prevalenceFigure7}
\alias{prevalenceFigure7}
\title{Figure 7. by year/month: plot for each database, diff lines per age group}
\usage{
prevalenceFigure7(prevalenceData)
}
\arguments{
\item{prevalenceData}{}
}
\value{
prevalenceFigure7
}
\description{
Figure 7. by year/month: plot for each database, diff lines per age group
}
