% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prevalenceFigures.R
\name{prevalenceFigure8}
\alias{prevalenceFigure8}
\title{Figure 8. by age group (x-axis) for databases (color) and sex (dashed/line)}
\usage{
prevalenceFigure8(prevalenceData)
}
\arguments{
\item{prevalenceData}{}
}
\value{
prevalenceFigure8
}
\description{
Figure 8. by age group (x-axis) for databases (color) and sex (dashed/line)
}
