% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardPlots.R
\name{prevalenceRatePerYearGroupBySexPlot}
\alias{prevalenceRatePerYearGroupBySexPlot}
\title{Prevalence Rate Per Year by Sex Plot}
\usage{
prevalenceRatePerYearGroupBySexPlot(prevalence_estimates)
}
\arguments{
\item{prevalence_estimates}{estimates of of the prevalence}
}
\description{
Prevalence Rate Per Year by Sex Plot
}
