% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardPlots.R
\name{prevalenceRatePerYearPlot}
\alias{prevalenceRatePerYearPlot}
\title{Prevalence Rate Per Year Plot}
\usage{
prevalenceRatePerYearPlot(prevalence_estimates)
}
\arguments{
\item{prevalence_estimates}{estimates of of the prevalence}
}
\description{
Prevalence Rate Per Year Plot
}
