% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prevalenceTables.R
\name{prevalenceTable3}
\alias{prevalenceTable3}
\title{Table 3. Number of cases and total number of patients per database}
\usage{
prevalenceTable3(prevalenceData)
}
\arguments{
\item{prevalenceData}{}
}
\value{
prevalenceTable3
}
\description{
Table 3. Number of cases and total number of patients per database
}
