% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataExtraction.R
\name{processCSV}
\alias{processCSV}
\title{`processCSV()` iterates every csv file and uses `loadFileData()` to check what type of result it is and adds it to a list}
\usage{
processCSV(csv_files)
}
\arguments{
\item{data}{Results}

\item{filesLocation}{A path for filelocation}

\item{configData}{Data from yaml configuration file}

\item{databaseName}{Database name from TreatmentPatterns}
}
\value{
A list with all the results organized by type
}
\description{
`processCSV()` iterates every csv file and uses `loadFileData()` to check what type of result it is and adds it to a list
}
