% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportGenerator.R
\name{reportGenerator}
\alias{reportGenerator}
\title{This application offers a visualization of the figures available to fill in a study report in Word format.}
\usage{
reportGenerator()
}
\value{
Dashboard
}
\description{
This application offers a visualization of the figures available to fill in a study report in Word format.
}
