% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportIncidencePrevalence.R
\name{reportIncidencePrevalence}
\alias{reportIncidencePrevalence}
\title{Creates an HTML report taking the prevalence and incidence estimates from IncidencePrevalence.}
\usage{
reportIncidencePrevalence(
  studyTitle,
  studyAuthor,
  prevalenceData,
  incidenceData
)
}
\arguments{
\item{studyTitle}{A character title of the study.}

\item{studyAuthor}{A character name of the author.}

\item{prevalenceData}{A tibble from IncidencePrevalence.}
}
\value{
An HTML document
}
\description{
Creates an HTML report taking the prevalence and incidence estimates from IncidencePrevalence.
}
