% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resultsDashboard.R
\name{resultDashboard}
\alias{resultDashboard}
\title{This application displays an automated and interactive report of the results from the IncidencePrevalence package.}
\usage{
resultDashboard(
  importFolderDenominator = here("inst/csv/denominatorMockData"),
  importFolderIndcidence = here("inst/csv/incidenceMockResults"),
  importFolderPrevalence = here("inst/csv/prevalenceMockResults")
)
}
\value{
Dashboard
}
\description{
This application displays an automated and interactive report of the results from the IncidencePrevalence package.
}
