% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treatmentPatterns.R
\name{saveAsFile}
\alias{saveAsFile}
\title{`saveAsFile()` saves an SVG-image from an html-file}
\usage{
saveAsFile(fileName, fileNameOut, zoom = 3, vwidth = 430, selector = NULL, ...)
}
\arguments{
\item{fileName}{HTML-Filename.}

\item{fileNameOut}{Filename of image, either with .pdf or .png extension.}

\item{zoom}{Zoom factor, default = 3.}

\item{vwidth}{Width of frame to capture, default = 430.}

\item{selector}{Default: `NULL`; May be set to `"svg"` only works if the output file is pdf, otherwise use `NULL`}

\item{...}{Other parameters to be passed to webshot2::webshot()}
}
\value{
Invisibly returns the normalized path to all screenshots taken. The character vector will have a class of '"webshot"'.
}
\description{
`saveAsFile()` saves an SVG-image from an html-file
}
\examples{
\dontrun{
  saveAsFile("webFile.html", "outputFile.png")}
}
