% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/studyReportApp.R
\name{studyReportApp}
\alias{studyReportApp}
\title{This application displays an automated and interactive report of the results from the IncidencePrevalence package.}
\usage{
studyReportApp()
}
\value{
An HTML document
}
\description{
This application displays an automated and interactive report of the results from the IncidencePrevalence package.
}
