% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/automaticText.R
\name{table1aAutText}
\alias{table1aAutText}
\title{Generates automatic text paragraph for Table 1}
\usage{
table1aAutText(attrition_data)
}
\arguments{
\item{incidence_attrition}{Data frame with incidence attrition data}

\item{prevalence_attrition}{Data frame with incidence attrition data}
}
\value{
Automatic text as a character string
}
\description{
Generates automatic text paragraph for Table 1
}
