% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardTables.R
\name{table1Att}
\alias{table1Att}
\title{`table1IncAtt()` creates table 1 with number of participants from each database from attrition data}
\usage{
table1Att(attritionData, attritionDataType)
}
\arguments{
\item{attritionDataType}{Either "incidence" or prevalence}

\item{incidence_attrition}{A dataframe with incidence attrition data.}
}
\description{
`table1IncAtt()` creates table 1 with number of participants from each database from attrition data
}
