% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardTables.R
\name{table1Inc}
\alias{table1Inc}
\title{`table1IncPrev()` returns Table 1 using incidence/prevalence estimates.}
\usage{
table1Inc(incidence_estimates)
}
\arguments{
\item{incidence_estimates}{A data frame with incidence attrition data.}

\item{prevalence_estimates}{A data frame with prevalence estimates.}
}
\description{
`table1IncPrev()` returns Table 1 using incidence/prevalence estimates.
}
