% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/incidenceText.R
\name{table1IncidenceParagraph}
\alias{table1IncidenceParagraph}
\title{Table 1 Paragraph. Incidence rate/s of drug/s use over calendar incidence_start_date (per month/year) overall}
\usage{
table1IncidenceParagraph(indcidenceData)
}
\value{
table1Paragraph
}
\description{
Table 1 Paragraph. Incidence rate/s of drug/s use over calendar incidence_start_date (per month/year) overall
}
