% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/previewTables.R
\name{table1NumParPreview}
\alias{table1NumParPreview}
\title{table1NumParPreview}
\usage{
table1NumParPreview(incidence_attrition, prevalence_attrition)
}
\arguments{
\item{incidence_attrition}{incidence of the attrition}

\item{prevalence_attrition}{prevalence of the attrition}
}
\description{
table1NumParPreview
}
