% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/incidenceTables.R
\name{table2Incidence}
\alias{table2Incidence}
\title{Table 2. Incidence data according to figures 1 and 2}
\usage{
table2Incidence(incidenceData)
}
\arguments{
\item{incidenceData}{}
}
\value{
table2Incidence
}
\description{
Table 2. Incidence data according to figures 1 and 2
}
