% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardTables.R
\name{table2IncOver}
\alias{table2IncOver}
\title{table2IncOver}
\usage{
table2IncOver(incidence_estimates)
}
\arguments{
\item{incidence_estimates}{estimates of of the incidence}
}
\description{
table2IncOver
}
\examples{
table2IncOver(incidence_estimates = ReportGenerator:::incidence_estimates_latest)
}
