% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardTables.R
\name{table4IncAge}
\alias{table4IncAge}
\title{table4IncAge}
\usage{
table4IncAge(incidence_estimates)
}
\arguments{
\item{incidence_estimates}{estimates of of the incidence}
}
\description{
table4IncAge
}
\examples{
table4IncAge(incidence_estimates = ReportGenerator:::incidence_estimates_latest)
}
