% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prevalenceTables.R
\name{table4Prevalence}
\alias{table4Prevalence}
\title{Table 4. Numeric values of displayed figures 5-8}
\usage{
table4Prevalence(prevalenceData)
}
\arguments{
\item{prevalenceData}{}
}
\value{
table4Prevalence
}
\description{
Table 4. Numeric values of displayed figures 5-8
}
