% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardTables.R
\name{table5IncSex}
\alias{table5IncSex}
\title{table5IncSex}
\usage{
table5IncSex(incidence_estimates)
}
\arguments{
\item{incidence_estimates}{estimates of of the incidence}
}
\description{
table5IncSex
}
