% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/automaticText.R
\name{tableAttrition}
\alias{tableAttrition}
\title{Generates automatic text paragraph for Table Incidence Attrition}
\usage{
tableAttrition(attrition_data)
}
\arguments{
\item{attritionData}{Data frame with incidence attrition data}
}
\value{
Automatic text as a character string
}
\description{
Generates automatic text paragraph for Table Incidence Attrition
}
