% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataExtraction.R
\name{unzipFiles}
\alias{unzipFiles}
\title{`unzipFiles()` to a folder per database in a temporary directory for later use in `exctractCSV()`}
\usage{
unzipFiles(unzipDir, fileDataPath)
}
\arguments{
\item{unzipDir}{Locations of the folder to unzip files}

\item{fileDataPath}{Locations of the files to unzip}
}
\value{
The databaseFolders to which the files were unzipped
}
\description{
`unzipFiles()` to a folder per database in a temporary directory for later use in `exctractCSV()`
}
