% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataExtraction.R
\name{variablesConfigAttrition}
\alias{variablesConfigAttrition}
\title{`variablesConfigAttrition()` extracts to variablesConfig file the columns to identify Cohort Attrition}
\usage{
variablesConfigAttrition(path)
}
\arguments{
\item{path}{To a cohort attrition CSV file}
}
\value{
Updates the yaml config file
}
\description{
`variablesConfigAttrition()` extracts to variablesConfig file the columns to identify Cohort Attrition
}
