% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataExtraction.R
\name{variablesConfigWriter}
\alias{variablesConfigWriter}
\title{Writes variablesConfig file}
\usage{
variablesConfigWriter(fileDataPath = NULL)
}
\arguments{
\item{fileDataPath}{A list of csv files}
}
\value{
Writes a csv file into the config folder of ReportGenerator
}
\description{
`variablesConfigWriter()` takes a list of csv files, extract their columns and writes de variables config file
}
