% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataExtraction.R
\name{variablesConfigYaml}
\alias{variablesConfigYaml}
\title{Writes variablesConfig file in Yaml}
\usage{
variablesConfigYaml(
  fileDataPath = NULL,
  version = "IncidencePrevalence v0.4.0"
)
}
\arguments{
\item{fileDataPath}{A list of csv files}

\item{version}{A string to identify which version of IncidencePrevalence is used to generate the results}
}
\value{
Writes a csv file into the config folder of ReportGenerator
}
\description{
`variablesConfigYaml()` takes a list of csv files, extract their columns and writes de variables config file
}
