% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/automaticText.R
\name{autoCaptionCharac}
\alias{autoCaptionCharac}
\title{Generic to create automatic text paragraph for the caption for table from summarised_characteristics object.}
\usage{
autoCaptionCharac(summarisedCharacteristics)
}
\arguments{
\item{uploadedFiles}{Data frame with summarised_characteristics data}
}
\value{
Automatic text as a character string
}
\description{
Generic to create automatic text paragraph for the caption for table from summarised_characteristics object.
}
