% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/previewTables.R
\name{createPreviewMenuTable}
\alias{createPreviewMenuTable}
\title{Create the table in which a user can choose the item for a preview.}
\usage{
createPreviewMenuTable(data)
}
\arguments{
\item{data}{the input data}
}
\description{
Create the table in which a user can choose the item for a preview.
}
