% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataExtraction.R
\name{denominatorExtraction}
\alias{denominatorExtraction}
\title{Extracts denominator data from files in CSV format.}
\usage{
denominatorExtraction(importFolder = here("inst/csv/denominatorMockData"))
}
\value{
A tibble
}
\description{
Extracts denominator data from files in CSV format.
}
