% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataExtraction.R
\name{extractCSV}
\alias{extractCSV}
\title{`extractCSV()`}
\usage{
extractCSV(databaseFolders, configData, logger)
}
\arguments{
\item{databaseFolders}{A list of full name folder locations}

\item{configData}{Configuration from yaml file}

\item{logger}{A logger object}
}
\value{
A list objects with summarisedResults
}
\description{
`extractCSV()`
}
