% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateMockData.R
\name{generateMockData}
\alias{generateMockData}
\title{Creates a mock data set for ReportGenerator}
\usage{
generateMockData(
  databaseName = c("CHUBX", "CPRD GOLD", "IMASIS", "IPCI", "SIDIAP"),
  simulatePopulation = TRUE,
  outputPath = file.path(getwd(), "results"),
  internal = FALSE
)
}
\arguments{
\item{databaseName}{A vector with the name in characters of each database.}

\item{simulatePopulation}{TRUE or FALSE to simulate different population sizes. TRUE is default.}

\item{outputDir}{A character vector of the directory to export mock data.}
}
\value{
csv files
}
\description{
`generateMockData()` uses [IncidencePrevalence::mockIncidencePrevalenceRef()] function to create csv files to test ReportGenerator.
}
