% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateMockData.R
\name{generateMockData}
\alias{generateMockData}
\title{Creates a mock data set for ReportGenerator}
\usage{
generateMockData(
  databaseName = c("CHUBX", "CPRD GOLD", "IMASIS", "IPCI", "SIDIAP"),
  simulatePopulation = FALSE
)
}
\arguments{
\item{databaseName}{A vector with the name in characters of each database.}
}
\value{
csv files
}
\description{
`generateMockData()` uses [IncidencePrevalence::mockIncidencePrevalenceRef()] function to create csv files to test ReportGenerator.
}
