% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateMockData.R
\name{generateMockData}
\alias{generateMockData}
\title{This function uses data mockIncidencePrevalenceRef function to simulate three databases in CSV format to test the ReportGenerator}
\usage{
generateMockData(databaseName = "IPCI")
}
\arguments{
\item{databaseName}{A vector that specifies the name of each database.}
}
\value{
CSV files inside inst/csv duckdb
}
\description{
This function uses data mockIncidencePrevalenceRef function to simulate three databases in CSV format to test the ReportGenerator
}
