% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateMockData.R
\name{generateMockData}
\alias{generateMockData}
\title{Creates a mock data set for ReportGenerator. The output is a zip folder for each database.}
\usage{
generateMockData(
  databaseName = c("CHUBX", "CPRD_GOLD", "IMASIS", "IPCI", "SIDIAP"),
  simulatePopulation = TRUE,
  outputPath = testthat::test_path("studies", "zip"),
  internal = FALSE
)
}
\arguments{
\item{databaseName}{A vector with the name in characters of each database.}

\item{simulatePopulation}{TRUE or FALSE to simulate different population sizes. TRUE is default.}

\item{outputPath}{A character vector of the path to export mock data.}

\item{internal}{interal usage}
}
\value{
csv files
}
\description{
`generateMockData()` uses functions such as [IncidencePrevalence::mockIncidencePrevalenceRef()] to create csv files to test ReportGenerator.
}
