% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateReport.R
\name{generateReport}
\alias{generateReport}
\title{Generate report}
\usage{
generateReport(reportDocx, dataReportList, fileName)
}
\arguments{
\item{reportDocx}{the report document}

\item{dataReportList}{the list of items to be added}

\item{fileName}{the name of the report file}
}
\value{
NULL, the report is written to given file
}
\description{
Generate report
}
