% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataExtraction.R
\name{incidenceExtraction}
\alias{incidenceExtraction}
\title{Extracts incidence data from files in CSV format and saves the result in RDS.}
\usage{
incidenceExtraction(
  importFolderIndcidence = here("inst/csv/incidenceMockResults"),
  studyName = "mock_data"
)
}
\arguments{
\item{importFolderIndcidence}{Location of incidence results.}

\item{studyName}{In a string, the name of the study.}
}
\value{
A tibble
}
\description{
Extracts incidence data from files in CSV format and saves the result in RDS.
}
