% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataExtraction.R
\name{incidenceExtraction}
\alias{incidenceExtraction}
\title{Extracts incidence data from files in CSV format.}
\usage{
incidenceExtraction()
}
\value{
A tibble
}
\description{
Extracts incidence data from files in CSV format.
}
