% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/incidencePrevalenceReport.R
\name{incidencePrevalenceReport}
\alias{incidencePrevalenceReport}
\title{Creates a written report from IncidencePrevalence.}
\usage{
incidencePrevalenceReport(
  title = NULL,
  authors = NULL,
  authorsInstitution = NULL,
  date = NULL,
  countries = NULL,
  abstractText = NULL,
  byCondition = TRUE,
  format = "word"
)
}
\arguments{
\item{title}{Title of the study in <chr>.}

\item{authors}{Name of the author in <chr>.}

\item{authorsInstitution}{Name of the institution performing the study in <chr>.}

\item{date}{Date relevant to the study in <chr>.}

\item{countries}{Country or countries that participated in the study.}

\item{abstractText}{Abstract in character.}

\item{byCondition}{TRUE if study is related to a condition, FALSE if it is related to drug utilisation.}

\item{format}{A string. Options are "word", "pdf" or "html".}
}
\value{
A WORD, PDF or HTML document.
}
\description{
Creates a written report from IncidencePrevalence.
}
