% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardPlots.R
\name{incidenceRatePerYearColorByAgePlot}
\alias{incidenceRatePerYearColorByAgePlot}
\title{incidenceRatePerYearColorByAgePlot}
\usage{
incidenceRatePerYearColorByAgePlot(incidence_estimates)
}
\arguments{
\item{incidence_estimates}{estimates of of the incidence}
}
\description{
incidenceRatePerYearColorByAgePlot
}
