% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardPlots.R
\name{incidenceRatePerYearFacetByDBAgeGroupPlot}
\alias{incidenceRatePerYearFacetByDBAgeGroupPlot}
\title{incidenceRatePerYearFacetByDBAgeGroupPlot}
\usage{
incidenceRatePerYearFacetByDBAgeGroupPlot(incidence_estimates)
}
\arguments{
\item{incidence_estimates}{estimates of of the incidence}
}
\description{
incidenceRatePerYearFacetByDBAgeGroupPlot
}
