% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardplots.R
\name{incidenceRatePerYearPlot}
\alias{incidenceRatePerYearPlot}
\title{incidenceRatePerYearPlot}
\usage{
incidenceRatePerYearPlot(incidence_estimates)
}
\arguments{
\item{incidence_estimates}{estimates of of the incidence}
}
\description{
incidenceRatePerYearPlot
}
