% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataExtraction.R
\name{joinDatabase}
\alias{joinDatabase}
\title{`joinDatabase()` joins several zip or csv folders into a list of dataframes.}
\usage{
joinDatabase(fileDataPath, fileName = NULL, csvLocation, logger)
}
\arguments{
\item{fileDataPath}{File path(s) in character}

\item{fileName}{Name of the file in character to process in case the input is only csv}

\item{csvLocation}{Path folder location to uncompress the zip files}

\item{logger}{logger object}
}
\value{
A list of dataframes
}
\description{
`joinDatabase()` joins several zip or csv folders into a list of dataframes.
}
