% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataExtraction.R
\name{joinDatabase}
\alias{joinDatabase}
\title{`joinDatabase()` gathers several zip or csv folders into a list of dataframes for ReportGenerator}
\usage{
joinDatabase(
  uploadedFiles = NULL,
  csvLocation,
  package = "IncidencePrevalence",
  versionData = "0.4.1"
)
}
\arguments{
\item{uploadedFiles}{A list of filepaths}

\item{csvLocation}{Temporary folder}

\item{package}{Name of the packages that generated the results}

\item{versionData}{Version of the package}
}
\value{
A list of dataframes
}
\description{
`joinDatabase()` gathers several zip or csv folders into a list of dataframes for ReportGenerator
}
