% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataExtraction.R
\name{joinDatabases}
\alias{joinDatabases}
\title{`joinDatabases()` joins several zip or csv folders into a list of dataframes.}
\usage{
joinDatabases(fileDataPath)
}
\arguments{
\item{fileDataPath}{List of full file locations}
}
\value{
A list of dataframes
}
\description{
`joinDatabases()` joins several zip or csv folders into a list of dataframes.
}
