% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataExtraction.R
\name{loadFileData}
\alias{loadFileData}
\title{Load file data and save it in a list.}
\usage{
loadFileData(
  data,
  fileName,
  configData,
  resultsData,
  resultsColumns,
  databaseName,
  logger
)
}
\arguments{
\item{data}{named list with data.}

\item{fileName}{the file to load}

\item{configData}{config loaded from file}

\item{resultsData}{the loaded file data}

\item{resultsColumns}{the loaded file columns}

\item{databaseName}{db name}

\item{logger}{logger object}
}
\value{
the list with data
}
\description{
Load file data and save it in a list.
}
