% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mockSampleCSV.R
\name{mockSampleCSV}
\alias{mockSampleCSV}
\title{Creates mock data that simulates an array of databases in CSV format to test the reportGenerator}
\usage{
mockSampleCSV(cdm = NULL, databaseName = c("Synthea", "IPCI", "CPRD"))
}
\arguments{
\item{cdm}{A cdm object from mockIncidencePrevalenceRef}

\item{databaseName}{A vector that specifies the name of each database.}
}
\value{
CSV files inside inst/csv
}
\description{
Creates mock data that simulates an array of databases in CSV format to test the reportGenerator
}
