% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataExtraction.R
\name{prevalenceExtraction}
\alias{prevalenceExtraction}
\title{Extracts prevalence data from files in CSV format nd saves the result in RDS.}
\usage{
prevalenceExtraction(
  importFolderPrevalence = here("inst/csv/prevalenceMockResults"),
  studyName = "mock_data"
)
}
\value{
A tibble
}
\description{
Extracts prevalence data from files in CSV format nd saves the result in RDS.
}
