% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataExtraction.R
\name{prevalenceExtraction}
\alias{prevalenceExtraction}
\title{Extracts prevalence data from files in CSV format.}
\usage{
prevalenceExtraction(
  importFolderPrevalence = here("inst/csv/prevalenceMockResults")
)
}
\value{
A tibble
}
\description{
Extracts prevalence data from files in CSV format.
}
