% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataExtraction.R
\name{prevalenceExtractionToRDS}
\alias{prevalenceExtractionToRDS}
\title{Extracts prevalence data from files in CSV format nd saves the result in RDS.}
\usage{
prevalenceExtractionToRDS(
  importFolderPrevalence = here("inst/csv/prevalenceMockResults"),
  studyName = "mock_data"
)
}
\arguments{
\item{importFolderPrevalence}{Location of prevalence results.}

\item{studyName}{In a string, the name of the study.}
}
\value{
A tibble
}
\description{
Extracts prevalence data from files in CSV format nd saves the result in RDS.
}
