% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardPlots.R
\name{prevalenceRatePerYearFacetByDBAgeGroupPlot}
\alias{prevalenceRatePerYearFacetByDBAgeGroupPlot}
\title{prevalenceRatePerYearFacetByDBAgeGroupPlot}
\usage{
prevalenceRatePerYearFacetByDBAgeGroupPlot(prevalence_estimates)
}
\arguments{
\item{prevalence_estimates}{estimates of of the prevalence}
}
\description{
prevalenceRatePerYearFacetByDBAgeGroupPlot
}
