% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardPlots.R
\name{prevalenceRatePerYearFacetByDBOutcomePlot}
\alias{prevalenceRatePerYearFacetByDBOutcomePlot}
\title{prevalenceRatePerYearFacetByDBOutcomePlot}
\usage{
prevalenceRatePerYearFacetByDBOutcomePlot(prevalence_estimates)
}
\arguments{
\item{prevalence_estimates}{estimates of of the prevalence}
}
\description{
prevalenceRatePerYearFacetByDBOutcomePlot
}
