% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportGenerator.R
\name{reportGenerator}
\alias{reportGenerator}
\title{ReportGenerator Shiny App}
\usage{
reportGenerator()
}
\description{
`ReportGenerator()` launches the package's main app. The user can upload a zip folder, and the function detects what figures and tables are available to generate a Word report.
}
\examples{
\donttest{
reportGenerator()
}
}
