% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportIncidencePrevalence.R
\name{reportIncidencePrevalence}
\alias{reportIncidencePrevalence}
\title{Creates an HTML report taking the prevalence and incidence estimates from IncidencePrevalence.}
\usage{
reportIncidencePrevalence(
  studyTitle,
  studyAuthor,
  denominatorData,
  prevalenceData,
  format = "word"
)
}
\arguments{
\item{studyTitle}{Title of the study.}

\item{studyAuthor}{Name of the author.}

\item{denominatorData}{A tibble from IncidencePrevalence.}

\item{prevalenceData}{A tibble from IncidencePrevalence.}

\item{format}{Output options are "word", "pdf" or "html".}
}
\value{
A WORD, PDF or HTML document.
}
\description{
Creates an HTML report taking the prevalence and incidence estimates from IncidencePrevalence.
}
