% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportIncidencePrevalence.R
\name{reportIncidencePrevalence}
\alias{reportIncidencePrevalence}
\title{Report generator for IncidencePrevalence package}
\usage{
reportIncidencePrevalence(x)
}
\arguments{
\item{x}{A tibble}
}
\value{
An HTML document
}
\description{
Report generator for IncidencePrevalence package
}
