% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportGeneratorApp.R, R/resultsDashboard.R
\name{resultsDashboard}
\alias{resultsDashboard}
\title{This application displays an automated and interactive report of the results from the IncidencePrevalence package.}
\usage{
resultsDashboard()

resultsDashboard()
}
\value{
Dashboard

Shiny app.
}
\description{
This application displays an automated and interactive report of the results from the IncidencePrevalence package.

Results Data Visualisation Dashboard
}
