% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardTables.R
\name{table1NumPar}
\alias{table1NumPar}
\title{table1NumPar}
\usage{
table1NumPar(prevalence_attrition, incidence_attrition)
}
\arguments{
\item{prevalence_attrition}{prevalence of the attrition}

\item{incidence_attrition}{incidence of the attrition}
}
\description{
table1NumPar
}
