% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataExtraction.R
\name{testData}
\alias{testData}
\title{`testData()` extracts data from zip results and saves it in .rda format}
\usage{
testData(testFilesIP = testthat::test_path("IncPrev", "0.6.0", "zip"))
}
\arguments{
\item{testFilesDir}{testthat dir location of the test files with version number of folder (internal use).}
}
\value{
sysdata.rda instruction
}
\description{
`testData()` extracts data from zip results and saves it in .rda format
}
