% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataExtraction.R
\name{testData}
\alias{testData}
\title{`testData()` extracts data from zip results and saves it in .rda format}
\usage{
testData()
}
\value{
sysdata.rda instruction
}
\description{
`testData()` extracts data from zip results and saves it in .rda format
}
