% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataExtraction.R
\name{testData}
\alias{testData}
\title{`testData()` extracts data from zip results and saves it in .rda format}
\usage{
testData(
  testFilesIP = testthat::test_path("IP", "0.6.0", "zip"),
  testFilesTP = testthat::test_path("TP", "2.5.2", "csv", "CHUBX"),
  testFilesPP = testthat::test_path("PP", "0.5.1", "zip"),
  testFilesCS = testthat::test_path("CS", "0.2.5", "zip"),
  logger
)
}
\arguments{
\item{testFilesIP}{testthat dir location of the test files with version number of folder (internal use).}

\item{testFilesTP}{testthat dir of treatment pattern files}

\item{testFilesPP}{testthat dir of patient profiles files}

\item{testFilesCS}{testthat dir of cohort survival files}

\item{logger}{logger object}
}
\value{
sysdata.rda instruction
}
\description{
`testData()` extracts data from zip results and saves it in .rda format
}
