% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/studyReportApp.R
\docType{data}
\name{ui}
\alias{ui}
\title{This application displays an automated and interactive report of the results from the IncidencePrevalence package.}
\format{
An object of class \code{shiny.tag.list} (inherits from \code{list}) of length 4.
}
\usage{
ui
}
\value{
An HTML document
}
\description{
This application displays an automated and interactive report of the results from the IncidencePrevalence package.
}
\keyword{datasets}
