% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataExtraction.R
\name{variablesConfigYaml}
\alias{variablesConfigYaml}
\title{Writes variablesConfig file in Yaml}
\usage{
variablesConfigYaml(
  fileDataPath = NULL,
  package = "IncidencePrevalence",
  version = NULL
)
}
\arguments{
\item{fileDataPath}{A list of zipped csv files.}

\item{package}{A string to identify the name of the package, default IncidencePrevalence.}

\item{version}{A string to identify which version of IncidencePrevalence is used to generate the results.}
}
\value{
Adds column names to the variablesConfig.yaml
}
\description{
`variablesConfigYaml()` the user can load the column names to the variablesConfig.yaml so ReportGenerator can recognize its data.
}
